moved tests to helpers_test unit test
